medicare/requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
# pymongo==4.4.1
python-dotenv==1.0.0